Rename work path to section
