+++ components/Section.tsx
@@ -1,7 +1,7 @@
 import { Done } from "@mui/icons-material";
 import { ReactNode } from "react";
 
-export function WorkPath({
+export function Section({
   children,
   done = false,
 }: {
@@ -11,12 +11,12 @@ export function WorkPath({
   return (
     <div className="border-gray-200 w-fit flex flex-col shrink-0">
       {children}
-      {done && <WorkPathDone />}
+      {done && <SectionDone />}
     </div>
   );
 }
 
-export function WorkPathHeader({
+export function SectionHeader({
   children,
   action,
 }: {
@@ -35,18 +35,18 @@ export function WorkPathHeader({
   );
 }
 
-export function WorkPathContent({ children }: { children: ReactNode }) {
+export function SectionContent({ children }: { children: ReactNode }) {
   return <div className="grow py-4">{children}</div>;
 }
 
-function WorkPathDone() {
+function SectionDone() {
   return (
     <div className="flex items-stretch">
       <div className="w-8 mr-4 py-4 flex justify-center">
         <Done className="text-green-600" />
       </div>
       <div className="grow mr-4 flex items-center text-gray-400">
-        Work path completed
+        Section completed
       </div>
     </div>
   );
